Convert everything to use 64 bit

+++ include/allocator.h
@@ -31,9 +31,9 @@
 
 
 // Calc number of blocks:
-#define INODE_BLOCKS (uint32_t)31250 // supports 1 million 128 byte INODES
-#define REF_BLOCKS  (uint32_t)(1 + (DISK_SIZE_IN_BLOCKS - 1 - INODE_BLOCKS) / BYTES_PER_BLOCK)
-#define DATA_BLOCKS (uint32_t)(DISK_SIZE_IN_BLOCKS - 1 - INODE_BLOCKS - REF_BLOCKS)
+#define INODE_BLOCKS (uint64_t)31250 // supports 1 million 128 byte INODES
+#define REF_BLOCKS  (uint64_t)(1 + (DISK_SIZE_IN_BLOCKS - 1 - INODE_BLOCKS) / BYTES_PER_BLOCK)
+#define DATA_BLOCKS (uint64_t)(DISK_SIZE_IN_BLOCKS - 1 - INODE_BLOCKS - REF_BLOCKS)
 
 
 // Layout
@@ -51,14 +51,14 @@ void init_allocator();
 
 
 // public: for data blocks: COPY ON WRITE
-int write_to_next_free_block(const uint8_t* buffer, uint32_t* block_number);
-int read_data_block(uint8_t* buffer, uint32_t block_number);
-int free_data_block(uint32_t block_number);
+int write_to_next_free_block(const uint8_t* buffer, uint64_t* block_number);
+int read_data_block(uint8_t* buffer, uint64_t block_number);
+int free_data_block(uint64_t block_number);
 
 
 // passthrough with bounds checking
-int read_inode_block(uint8_t* buffer, uint32_t inode_block_number);
-int write_inode_block(const uint8_t* buffer, uint32_t inode_block_number);
+int read_inode_block(uint8_t* buffer, uint64_t inode_block_number);
+int write_inode_block(const uint8_t* buffer, uint64_t inode_block_number);
 
 // for formatting
 int format_super_block();
